+++ CHANGELOG.md
@@ -70,6 +70,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Graphviz always crash with this simple dot file #167
 - Seg fault in dot #1771
 - gml2gv doesn't handle some attributes correctly #1869
+- Add missing circo, fdp, neato, osage, patchwork, sfdp & twopi tools to Windows builds (copies of dot)
+- Add gv2gml tool to CMake (copy of gml2gv on Windows, symlink to gml2gv otherwise)
 
 ## [2.44.1] - 2020-06-29